// migrations/snapshots/5e176360572bad9d23b30c4b48e840aa6d0f7dfcdead4bf17af4ff8ccc9df0b5/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly priority: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly subject: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly priority: CodecTypes['pg/text@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly subject: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly priority: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly subject: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly priority: CodecTypes['pg/text@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly subject: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;