// migrations/snapshots/c0ebe0460e36c79aa2217193f60ce7c9a95143c1790b139d2ffcd7211a6ebd41/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly priority: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly subject: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly priority: CodecTypes['pg/text@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly subject: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly priority: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly subject: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly priority: CodecTypes['pg/text@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly subject: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;